친구요청 수락 로직 완성

+++ packages/server/src/users/users.service.ts
@@ -10,7 +10,7 @@ import {
 import { Server, Socket } from 'socket.io';
 import { GetUserRoomDto, GetUserRoomsDto } from 'src/chats/dto/rooms.dto';
 import { ChannelParticipant, DmParticipant } from 'src/chats/rooms.entity';
-import { EntityNotFoundError, Like, Not, Repository } from 'typeorm';
+import { Like, Not, Repository } from 'typeorm';
 import { AuthUserDto } from './dto/auth-user.dto';
 import { CreateUserDto } from './dto/create-user.dto';
 import { PatchNicknameDto } from './dto/patch-user.dto';
@@ -30,11 +30,12 @@ import {
   IsBlockedDto,
 } from './dto/user.dto';
 import { v4 as uuidv4 } from 'uuid';
-import { Auth, Block, Friend, Request, User, Rank } from './users.entity';
+import { Auth, Block, Friend, Request, User, Rank, Game } from './users.entity';
 import { ConnectionDto, ConnectionsDto } from './dto/connect-user.dto';
 import { WsException } from '@nestjs/websockets';
 import { MailerService } from '@nestjs-modules/mailer';
 import * as AWS from 'aws-sdk';
+import { GameLadderDto, GameRecordDto, WatchDto } from 'src/game/game.dto';
 
 @Injectable()
 export class UserService {
@@ -57,6 +58,8 @@ export class UserService {
     private channelParticipantsRepository: Repository<ChannelParticipant>,
     @Inject('DMPARTICIPANTS_REPOSITORY')
     private dmParticipantsRepository: Repository<DmParticipant>,
+    @Inject('GAMES_REPOSITORY')
+    private gamesRepository: Repository<Game>,
     @Inject('RANK_REPOSITORY')
     private rankRepository: Repository<Rank>,
     private readonly mailerService: MailerService,
@@ -110,9 +113,8 @@ export class UserService {
 
   async getUser(id: number): Promise<UserDto> {
     const user = await this.usersRepository.findOne({ where: [{ id: id }] });
-    // if (user === null) throw new EntityNotFoundError(User, id);
 
-    if (!user) return null;
+    if (!user) throw new NotFoundException(`User with id ${id} not found`);
     const userDto = new UserDto(user.id, user.nickname, user.image);
     return userDto;
   }
@@ -200,7 +202,8 @@ export class UserService {
       relations: { user: true },
       where: { user: { id: id } },
     });
-    // if (auth === null) throw new EntityNotFoundError(Auth, id);
+    if (auth === null)
+      throw new NotFoundException('유저의 인증 정보가 없습니다.');
 
     const authUserDto = new AuthUserDto();
     authUserDto.authenticated = auth.authenticated;
@@ -241,7 +244,8 @@ export class UserService {
     const blockedUser = await this.usersRepository.findOne({
       where: [{ id: userId }],
     });
-    if (blockedUser === null) throw new EntityNotFoundError(User, id);
+    if (blockedUser === null)
+      throw new NotFoundException('User who blocked not found');
 
     // FIXME: 테스트용
     let blockId = 1;
@@ -267,6 +271,7 @@ export class UserService {
       relations: ['user', 'blockedUser'],
       where: [{ user: { id: id }, blockedUser: { id: userId } }],
     });
+    if (blockRow === null) throw new NotFoundException('Block not found');
 
     this.blocksRepository.delete(blockRow.id);
     const user: UserDto = new UserDto(
@@ -419,7 +424,7 @@ export class UserService {
     const friend = await this.usersRepository.findOne({
       where: [{ id: friendId }],
     });
-    if (friend === null) throw new EntityNotFoundError(User, friendId);
+    if (friend === null) throw new NotFoundException('없는 유저입니다.');
 
     let rowId = 1;
     const maxFriendId = await this.friendsRepository
@@ -462,6 +467,8 @@ export class UserService {
   }
 
   async getFriends(id: number): Promise<FriendDto[]> {
+    const user = await this.usersRepository.findOne({ where: [{ id: id }] });
+    if (user === null) throw new NotFoundException('유저가 존재하지 않습니다.');
     // SELECT * FROM public."friend"
     // LEFT JOIN "user" ON "user"."id" = friendId
     // WHERE "friend"."userId = "user"."id";
@@ -504,7 +511,8 @@ export class UserService {
     const responser = await this.usersRepository.findOne({
       where: [{ id: responserId }],
     });
-    if (responser === null) throw new EntityNotFoundError(User, responserId);
+    if (responser === null)
+      throw new NotFoundException('유저가 존재하지 않습니다.');
 
     const duplicateRequestCheck = await this.requestsRepository.findOne({
       where: { requestor: { id: requestorId }, responser: { id: responserId } },
@@ -556,6 +564,33 @@ export class UserService {
     });
 
     return requestDtoList;
+  }
+
+  async getLadderStat(id: number): Promise<GameLadderDto> {
+    const rank = await this.rankRepository.findOne({
+      relations: { user: true },
+      where: { user: { id: id } },
+    });
+
+    if (rank === null) throw new NotFoundException('랭크 정보가 없습니다.');
+
+    const ladderStat = new GameLadderDto(rank);
+    return ladderStat;
+  }
+
+  async getGameRecords(id: number): Promise<GameRecordDto[]> {
+    const gameRecords = await this.gamesRepository.find({
+      relations: { user: true, opponent: true },
+      where: { user: { id: id } },
+    });
+
+    const gameRecordDtoList: GameRecordDto[] = [];
+    gameRecords.map((gameRecord) => {
+      const gameRecordDto = new GameRecordDto(gameRecord);
+      gameRecordDtoList.push(gameRecordDto);
+    });
+
+    return gameRecordDtoList;
   }
 
   // ANCHOR: Socket
